fastapi-version/requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/dev.in
#
annotated-types==0.7.0
    # via
    #   -r requirements/prod.txt
    #   pydantic
anyio==4.6.2.post1
    # via
    #   -r requirements/prod.txt
    #   starlette
black==24.10.0
    # via -r requirements/dev.in
build==1.2.2.post1
    # via pip-tools
cassandra-driver==3.29.2
    # via -r requirements/prod.txt
click==8.1.7
    # via
    #   -r requirements/prod.txt
    #   black
    #   geomet
    #   pip-tools
coverage==7.6.4
    # via -r requirements/dev.in
fastapi==0.115.3
    # via -r requirements/prod.txt
flake8==7.1.1
    # via -r requirements/dev.in
geomet==0.2.1.post1
    # via
    #   -r requirements/prod.txt
    #   cassandra-driver
gevent==24.10.3
    # via -r requirements/prod.txt
greenlet==3.1.1
    # via
    #   -r requirements/prod.txt
    #   gevent
gunicorn==23.0.0
    # via -r requirements/prod.txt
idna==3.10
    # via
    #   -r requirements/prod.txt
    #   anyio
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==24.1
    # via
    #   -r requirements/prod.txt
    #   black
    #   build
    #   gunicorn
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r requirements/dev.in
platformdirs==4.3.6
    # via black
pycodestyle==2.12.1
    # via flake8
pydantic==2.9.2
    # via
    #   -r requirements/prod.txt
    #   fastapi
pydantic-core==2.23.4
    # via
    #   -r requirements/prod.txt
    #   pydantic
pyflakes==3.2.0
    # via flake8
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
redis==5.2.0
    # via -r requirements/prod.txt
six==1.16.0
    # via
    #   -r requirements/prod.txt
    #   geomet
sniffio==1.3.1
    # via
    #   -r requirements/prod.txt
    #   anyio
starlette==0.41.0
    # via
    #   -r requirements/prod.txt
    #   fastapi
typing-extensions==4.12.2
    # via
    #   -r requirements/prod.txt
    #   fastapi
    #   pydantic
    #   pydantic-core
wheel==0.44.0
    # via pip-tools
zope-event==5.0
    # via
    #   -r requirements/prod.txt
    #   gevent
zope-interface==7.1.1
    # via
    #   -r requirements/prod.txt
    #   gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
